tests: Moving logging test to separate file
